Weather-App/requirements.txt
PyQt5
QT-PyQt-PySide-Custom-Widgets
requests
mysql.connector